Moved the db_model code to deps
